Normative: Clarify sign of modulo operation in DurationHandleFractions
Unlike the `%` operator, the result of the modulo operation in the spec
text has the sign of the right-hand operand, not the left-hand one. So
the current description of DurationHandleFractions was incorrect, and
would lead to throwing due to mixed signs in CreateTemporalDuration.

Add tests that will throw if this is implemented incorrectly.