Normative: Duration operations should be the same in PlainDate.add and Calendar.dateAdd
Fixes a discrepancy between how durations are treated depending on whether
they are passed to PlainDate.prototype.add() or
Calendar.prototype.dateAdd(). The treatment should be the same in both
cases.

In the case of PlainDate.prototype.subtract(), the duration should be
negated before being passed to Calendar.prototype.dateAdd().

Closes: #1708