Normative: Consistent order of operations in PMD.toPD and PYM.toPD
The spec text was inconsistent about whether the type check on the
argument or the calendar.fields checks should occur first. In
Temporal.PlainMonthDay.prototype.toPlainDate, the fields checks were
first. In Temporal.PlainYearMonth.prototype.toPlainDate, the type check
was first. This changes the PlainMonthDay algorithm to match the
PlainYearMonth one, and adds tests to both.